pyproject.toml
[project]
name = "LLMover"
version = "0.1.0"
description = "A tool to move LLM models from local storage to USB stick with symlink creation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
llm-mover = "llm_mover.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_mover"]